Разбивает utils.js на несколько файлов
