+++ src/view/point-edit.js
@@ -1,6 +1,6 @@
 import dayjs from 'dayjs';
 import {DESTINATIONS} from '../const.js';
-import {getRandomInteger} from '../utils.js';
+import {getRandomInteger} from '../utils/common.js';
 import AbstractView from './abstract.js';
 
 function createDestinationDatalistTemplate(destinations) {
